[Backend] Actualizacion de los Crud de arrendatario,pagos,mantenimienti,,propietarios,contratos

+++ database/migrations/2024_05_16_021447_create_mantenimientos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('propiedad_id')->references('id')->on('propiedades')->onDelete('cascade');
             $table->text('descripcion');
             $table->date('fecha');
-            $table->decimal('costo', 10, 2);
+            $table->bigInteger('costo');
             $table->timestamps();
         });
     }
